pyproject.toml
[tool.poetry]
name = "crpt-test-assignment-2"
version = "0.1.0"
description = ""
authors = ["Vyacheslav Ratushnyi <ratushnyyvm@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.7"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
